Switch to letting the connector manage offsets
Prior to this commit, the Connect framework would periodically call
flush() on the OSSinkTask and assume that when this returned it had
successfully uploaded all of the SyncRecords to Object Storage. Now
the OSSinkTask tracks which topic partitions it has been assigned and
reports the offset that Kafka can safely commit via the preCommit()
method.

The intent behind this change is to pave the way for future updates to
the connector which will allow it to batch up the data from multiple
SyncRecords into a single COS object.

This changes is quite large because it contains a couple of re-factors:
  - Delegate writes to object storage to a PartitionWriter class. The
    OSSinkTask tracks one instance of PartitionWriter per Kafka
    partition assigned to it. In the future, the implementation of
    PartitionWriter will contain the batching logic.
  - Create a Bucket class as part of the abstraction layer in-front of
    the AmazonS3 COS implementation. This binds together an instance of
    the AmazonS3 API and a bucket name, so that bucket name doesn't need
    to be propagated all over the connector code.