Add support for writing objects after a deadline
Add the 'os.object.deadline.seconds' property which when set to a
non-zero value is the number of seconds after which all of the Kafka
records received so far will be written into an object storage object.

This time interval is measured based on the wall clock of the connector.

From an implementation perspective, the deadline is handled via a
`ScheduledExecutorService` wrapped into `DeadlineServiceImpl`. This
introduces the potential of concurrency into `OSPartitionWriter`, which
previously was only accessed by a single thread.

To minimise the complexity of this added concurrency, the `RequestProcessor`
class has been added. This class maintains a queue of events and
serializes their delivery.