Create records based on Kafka timestamp interval
Add an `os.object.interval.seconds` property which configures an
interval used to determine which Kafka records are grouped together into
an object. If the connector has not started grouping together records,
then the Kafka timestamp of the next Kafka record it receives is used as
the starting time of the interval. The connector will then group
together records until it receives a record that has a timestamp that
exceeds the interval value. This record becomes the first record in the
next group, and so on.