Refactor: simplify object completion code
Prior to this re-factor the logic for deciding when to write objects to
object storage was spread across several places in the code base.
This made the code difficult to extend with different schemes for deciding when
to write objects, and difficult to test - as large amounts of code needed to be
tested at once.

This commit adds a `ObjectCompletionCriteria` interface which abstracts
the logic for deciding when to write an object, and a
`CompletionCriteriaSet` which groups together instances of this
interface. The set is then used by `OSPartitionWriter` to decide when to
write objects into object storage.